Add files via upload
The purpose of the code was to demonstrate that certain algorithms are embarrassingly parallel. Specifically, the task was to take a sequential program that searches for the global maximum (mana level) within a defined grid containing multiple local maxima (peaks) and then implement its parallel version to show the performance benefits of parallelization.